link/prolink: relocate the Pioneer wire encoders out of analysis (part 3c)
Move the Pro DJ Link waveform/PQT2/PSSI/PVB2/PVBR generators and the
pioneerEncoder into a new link/prolink package, leaving analysis as DSP plus
the neutral core.Analysis projection. AnalyzeTrack delegates to the installed
Encoder (main installs prolink.NewEncoder() at startup); the direct callers in
dbserver/api/tools switch to prolink. The encoder tests move with the code
(they'd cycle if left in package analysis).

Kept in analysis where a non-serving consumer needs them without inverting the
dependency: the beat-grid encoders (GenerateBeatGrid[FromBeats] — used by the
ANLZ import path, so prolink calls them via analysis and no logic is
duplicated) and the ANLZ container / VBR-seek helpers (WrapANLZ, VBRSeekIndex,
section IO — used by dbserver/export/import). Those relocate later, once the
rekordbox import/export is extracted.

Byte-safe: all 10 colour-waveform + blob golden hashes unchanged; analysis,
prolink, api, and dbserver suites pass. analysisRate is exported as
AnalysisRate so the relocated tests share the DSP rate.
